Change to auth method
As discussed in the technical working group (https://groups.google.com/forum/#!topic/matchmaker-exchange-technical/Wf-TwRcAUnc) it would be better to have the auth token in the header of the request as opposed to the query string. It was agreed this should also effect v1, as the v1 document doesn't specify this directly.

I have made this as a pull request to allow others to comment in case it was missed in the google group. (This is how the Decipher implementation will be delivered)